Refactor MachineController
- Remove `IdleController` and move implementation into `MachineController`
- Inject `ShellController` instead of creating multiple
